agent/requirements.txt
pyyaml
requests
colorama
